requirements.txt
azure-storage-blob==12.23.1
beautifulsoup4==4.12.3
Faker==30.3.0
numpy==2.1.2
pandas==2.2.3
psycopg2_binary==2.9.9
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.39.0
